buffer: force flushing on buffer full
There is a static buffer allocated per stream, which is by default set
to 4 KB, this buffer is protected to not overflow but not checked if
full, in this case, the program is just ignoring new data infinitely and
never fetch anything new not push anything.

This commit fix this by forcing flush on buffer full.